+++ backend/src/baserow/core/registries.py
@@ -1,6 +1,11 @@
 from .registry import (
-    Instance, Registry, ModelInstanceMixin, ModelRegistryMixin, APIUrlsRegistryMixin,
-    APIUrlsInstanceMixin
+    Instance,
+    Registry,
+    ModelInstanceMixin,
+    ModelRegistryMixin,
+    APIUrlsRegistryMixin,
+    APIUrlsInstanceMixin,
+    ImportExportMixin,
 )
 from .exceptions import ApplicationTypeAlreadyRegistered, ApplicationTypeDoesNotExist
 
@@ -89,7 +94,7 @@ class PluginRegistry(APIUrlsRegistryMixin, Registry):
     register extra api and root urls.
     """
 
-    name = 'plugin'
+    name = "plugin"
 
     @property
     def urls(self):
@@ -107,7 +112,9 @@ class PluginRegistry(APIUrlsRegistryMixin, Registry):
         return urls
 
 
-class ApplicationType(APIUrlsInstanceMixin, ModelInstanceMixin, Instance):
+class ApplicationType(
+    APIUrlsInstanceMixin, ModelInstanceMixin, ImportExportMixin, Instance
+):
     """
     This abstract class represents a custom application that can be added to the
     application registry. It must be extended so customisation can be done. Each
@@ -150,6 +157,52 @@ class ApplicationType(APIUrlsInstanceMixin, ModelInstanceMixin, Instance):
         :type application: Application
         """
 
+    def export_serialized(self, application):
+        """
+        Exports the application to a serialized dict that can be imported by the
+        `import_serialized` method. The dict is JSON serializable.
+
+        :param application: The application that must be exported.
+        :type application: Application
+        :return: The exported and serialized application.
+        :rtype: dict
+        """
+
+        return {
+            "id": application.id,
+            "name": application.name,
+            "order": application.order,
+            "type": self.type,
+        }
+
+    def import_serialized(self, group, serialized_values, id_mapping):
+        """
+        Imports the exported serialized application by the `export_serialized` as a new
+        application to a group.
+
+        :param group: The group that the application must be added to.
+        :type group: Group
+        :param serialized_values: The exported serialized values by the
+            `export_serialized` method.
+        :type serialized_values: dict`
+        :param id_mapping: The map of exported ids to newly created ids that must be
+            updated when a new instance has been created.
+        :type id_mapping: dict
+        :return: The newly created application.
+        :rtype: Application
+        """
+
+        if "applications" not in id_mapping:
+            id_mapping["applications"] = {}
+
+        serialized_copy = serialized_values.copy()
+        application_id = serialized_copy.pop("id")
+        serialized_copy.pop("type")
+        application = self.model_class.objects.create(group=group, **serialized_copy)
+        id_mapping["applications"][application_id] = application.id
+
+        return application
+
 
 class ApplicationTypeRegistry(APIUrlsRegistryMixin, ModelRegistryMixin, Registry):
     """
@@ -159,7 +212,7 @@ class ApplicationTypeRegistry(APIUrlsRegistryMixin, ModelRegistryMixin, Registry
     register api related urls.
     """
 
-    name = 'application'
+    name = "application"
     does_not_exist_exception_class = ApplicationTypeDoesNotExist
     already_registered_exception_class = ApplicationTypeAlreadyRegistered