+++ backend/tests/fixtures/__init__.py
@@ -9,9 +9,19 @@ from .table import TableFixtures
 from .view import ViewFixtures
 from .field import FieldFixtures
 from .token import TokenFixtures
+from .template import TemplateFixtures
 
 
-class Fixtures(SettingsFixtures, UserFixtures, UserFileFixtures, GroupFixtures,
-               ApplicationFixtures, TableFixtures, ViewFixtures, FieldFixtures,
-               TokenFixtures):
+class Fixtures(
+    SettingsFixtures,
+    UserFixtures,
+    UserFileFixtures,
+    GroupFixtures,
+    ApplicationFixtures,
+    TableFixtures,
+    ViewFixtures,
+    FieldFixtures,
+    TokenFixtures,
+    TemplateFixtures,
+):
     fake = Faker()